Split Raster into Dataset & Raster
Currently, Raster instances may contain one or more bands. As a result,
ISCE functions that want to operate on a single band must either check
that the input Raster contains a single band or else take the band index
as a separate argument. This potentially leads to a lot of bloat in
error checking and long function signatures.

This commit attempts to solve this by splitting Raster into two separate
classes: Dataset and Raster. Datasets are a collection of one or more
Raster bands (roughly equivalent to a GDALDataset). Rasters always
correspond to a single band (similar to GDALRasterBand).

Rough summary of changes:
 - Separate Raster into two classes, Dataset and Raster
 - Raster & Dataset placed under isce::io::gdal namespace (to avoid
   possible future name collisions with other types of dataset)
 - Add constructors for in-memory Raster/Dataset from pointer & sizes (&
   strides) - read-only access if pointer is const, defaults to
   read/write access otherwise)
 - Improve error checking in GDAL interface (e.g. writing to read-only
   Raster is no longer allowed, exceptions are thrown on invalid EPSG
   codes)
 - Raise error if geotransform orientation is not aligned with
   projection system axes
 - Label const methods as const! (const Raster implies read-only)
 - Change Raster dimensions from size_t -> int
 - Allow compile-time conversion from template type to GDALDataType for
   Raster I/O (previously implemented with std::unordered_map)
 - Rename Raster I/O methods
 - Change Raster::getEPSG()/setEPSG() ->
   Raster::getProjection()/setProjection()
 - Add isce::except::OutOfRange and isce::except::GDALError exceptions
 - Use OGRSpatialReference::AutoIdentifyEPSG() rather than
   ""::FindMatches() to identify missing EPSG code as it performs better