Refactor CI builds (#289)
* Refactor CI builds

* Generalize script to allow CI image suites

For example, change IMAGE_SUITE to 'conda' or 'ubuntu'
in .ci/run_docker_tests.sh

* Use ubuntu images for testing

The conda images aren't passing tests on CI currently
Let's just see what happens when we build on ubuntu instead

* Revert "Use ubuntu images for testing"

This reverts commit 9fb833bb418d6d2c4518594b9b40e6cb2ac92051.

* Activate conda base env in testing images

This is needed for PROJ data to load properly. Still not sure
exactly why, but tested at it fixes the previous CI issues.

https://github.com/conda-forge/proj.4-feedstock/issues/49

* Activate conda via .bashrc

Unfortunately docker commands are running as pure sh and don't understand
the conda activation. This is how the official conda images do it.
https://github.com/ContinuumIO/docker-images

* Activate conda in the run command itself

* Run bash interactively for tests

This is so that .bashrc is loaded and the conda env is set up properly