Auto-enable cuda if found
This simplifies build instructions and is more reasonable default behavior IMO.
If cmake can find cuda itself, cuda functionality is automatically turned on.

Explicitly setting the old `WITH_CUDA` option can still be used to force
cuda features ON or OFF. If not set, WITH_CUDA defaults to `Auto` which
will auto-detect the cuda toolchain and use it if possible.