Add legacy Raster constructor from isce::io::gdal::Raster (#407)
Allows converting a new-style Raster to the legacy class. The
constructed Raster object is non-owning and tied to the lifetime of the
original Raster object.

Attempting to copy a non-owning Raster throws an exception to guard
against code that expects the Raster to persist beyond the lifetime of
the calling scope.

Also, fixes a bug in isce::io::Raster's assignment operator which failed
to clean up the existing _dataset member.