Style updates to some cuda sources (#469)
* Fix include ordering of .cu files

clang-format doesn't recognize .cu files as "main" in
file-to-main-include mappings by default. As a result, the corresponding
header files for some CUDA source files might not be included as the
first include.

e.g. "gpuLUT2d.cu" should have "gpuLUT2d.h" as its first include but
clang-format was putting it last

This commit adds .cu as a valid source file extension for performing
this matching.

* Style changes to gpuLUT2d

* Style changes to gpuGeometry