GPU azimuth compression (#475)
* Add (cuda) RadarGeometry

- Add copy constructor and interp method accessor for gpuLUT2d

* Add error code value for rdr2geo/geo2rdr failure to converge

* Add unified host/device rdr2geo implementation

* Add unified host/device geo2rdr implementation

(for 2D Doppler models only)

* Allow constructing gpuDEMInterpolator from const DEMInterpolator

* Fix(?) bug causing GPU out-of-memory on gpuDEMInterpolator construction

If the input (host) DEMInterpolator did not contain Raster data, its
length and width are never set, and therefore have indeterminate values.
gpuDEMInterpolator's constructor from host DEMInterpolator was then
allocating GPU memory according to the dims of its input, sometimes
resulting in arbitrarily large memory allocation.

* Add prelim GPU backprojection code

Just checkpointing - some stuff still missing

* Add helpers for converting host Kernels to device Kernels

- Add KernelType enum type so the derived type of polymorphic Kernel
  objects can be identified
- Add constructors for each device Kernel from the corresponding host
  Kernel
- Changed TabulatedKernel's & ChebyKernel's underlying storage type from
  valarray to vector to facilitate copying to the device
- Add public methods exposing TabulatedKernel's table and ChebyKernel's
  coeffs

* Finish GPU backprojection implementation

- Add conversion from polymorphic host Kernel to static device Kernel
- Compute min/max coherent processing window bounds from among all
  targets
- Add batch processing loop over range-compressed pulses within coherent
  processing bounds
- Add coherent summation implementation

* Add (broken) pybind11 bindings for GPU backproject

Adding these bindings causes the pybind11 library to generate runtime
errors (undefined symbol). Not sure why this is happening. The offending
lines are commented out for now.

* Fix typo in pybind11 module name

This caused the wrong version of "backproject" to be run by tests

* Fix pybind11 extension linking to libiscecuda

The cmake variable LISCECUDA was not defined at the scope at which it
was linking to the extension library so cmake was quietly ignoring it

- Uncomment cuda backprojection bindings

* Add unit test for isce.cuda.focus.backproject

Mostly copy-paste from the CPU backprojection unit test. Not sure how to
avoid this code duplication within the current unit testing system.

* Fix bug with pybind_isce3 cuda module

We were trying to build CUDA components of the pybind11 extension module
without first checking whether CUDA was enabled. This commit adds guards
to make sure the pybind11 bindings are safe to build without CUDA
enabled.

- Define a preprocessor macro "ISCE3_CUDA" in c++ code if CUDA
  support has been enabled
- Conditionally generated the pybind_isce3.cuda submodule depending on
  the value of this macro

* Add support for 'gpu_enabled' flag in RSLC workflow

- If specified azimuth compression block width is zero or negative, make
  the block span the full swath width. This is often the most efficient
  way of tiling, but the memory footprint is unpredictable

- Increase default azimuth compression block size from 64x64 to
  2048x2048. Larger block sizes are more efficient in general. The chosen
  default keeps GPU memory footprint around 1 GB

* Fixup docstrings

- Fix typo and copy-paste error in various docstrings
- Remove resolved comment

* Use typeid like a real c++ programmer

Remove KernelType enumeration used to determine runtime type of a Kernel
class and use the builtin typeid operator for this instead

* Use a less ambiguous variable name for dry tropo model

"dtm" could be confused with Digital Terrain Model

* Set default gpu_enabled flag to null

This makes the RSLC workflow automatically use GPU processing if the
isce3 library was configured with CUDA enabled.

It also allows user runconfigs to do the same (previously, if
gpu_enabled was unspecified, it would revert to the default)

* Halt execution if GPU was requested but not available

* Downgrade missing attitude log message to warning

Was previously an error, but processing would continue, falling back to
a zero-Doppler assumption