Remove gpuPolarBenchmark unit test (#490)
This appears to be a microbenchmark rather than a unit test. It is never
run as part of our unit testing suite and I wasn't previously aware of
it.

It no longer builds correctly (if it ever did in the first place) and
isn't up-to-par with our current coding standards. There seem to be other
existing tests exercising the same functionality.