Resolve warnings about unknown openmp pragmas (#590)
OpenMP is a (optional) private dependency of libisce3, meaning that
downstream build targets that depend on isce3 do not inherit this
dependency. However, we include some OpenMP pragmas in header and
inline implementation files. This should be considered benign - code
with such pragmas is still valid when compiled without OpenMP - but it
causes some compiler warnings to be generated when building the test
suite with '-Wunknown-pragmas'.

This update adds OpenMP as a (optional) dependency of each C++/CUDA test
in order to avoid these warnings.