Fix geo2rdr segmentation fault (#587)
* Fix geo2rdr segmentation fault

Add bounds checking logic to geo2rdr in order to avoid evaluating a
Doppler LUT outside of its valid bounds. Internally, LUT2d::eval()
writes to a pyre error channel, which may cause a segmentation fault if
executed in multi-threaded code.

As a consequence of this change, if an out-of-bounds lookup is
encountered during geo2rdr, it is treated as a failure to converge. This
isn't an ideal error-handling mechanism since it obscures the cause of
the failure, but more robust error-handling would require significant
refactoring.

* Update missing case handling

- Remove default case so that a missing case label will trigger a
  warning from -Wswitch
- Turn -Wswitch warnings into errors