Provide CMake version info to C++/Python components (#608)
* Add version setup logic to CMake

Version info is loaded from a VERSION.txt file in the project root.
The file is of the form e.g. "0.1.0+dev", that is, a conventional
3-component semantic version followed by an optional descriptor which
provides more info about the state of the source tree.

The special descriptor "dev" is automatically replaced with the git hash
if possible (and otherwise will be left as is, e.g. if git info is not
available).

When updating isce3, the proposed strategy is to first add a commit
to bump the VERSION.txt file to the new release version,
e.g. "0.2.0+release", and then after tagging that release, immediately
updating the descriptor in a subsequent commit, e.g. "0.2.0+dev".
This will make it easy to determine when the version of isce3 in use is
not an explicitly tagged release.

* Add C++ config header

* Add version/config info to pybind11 extension module

* Specify linguist language for template files

Co-authored-by: Geoffrey M Gunter <geoffrey.m.gunter@jpl.nasa.gov>

* Streamline version info

* Pybind info reduced to full version string in __version__
* Added optional prerelease segment
  (https://www.python.org/dev/peps/pep-0440/#pre-release-separators)
* Replace extra isce3 version macros with single extern string variable

Co-authored-by: Geoffrey M Gunter <geoffrey.m.gunter@jpl.nasa.gov>

* Remove old code from pybind test

Co-authored-by: Geoffrey M Gunter <geoffrey.m.gunter@jpl.nasa.gov>