Patch DenseMatrix to fix matrix multiplication on CUDA 11.3 (#809)
* Patch DenseMatrix to fix matrix multiplication on CUDA 11.3

This update patches the DenseMatrix class to workaround an issue with
recent releases of Eigen and some CUDA configurations.

When compiled with nvcc 11.3, Eigen's matrix-matrix and matrix-vector
product operations compile to nothing (no assembly instructions are
emitted). In addition, these operations result in "illegal memory
access" errors when compiled with nvcc 11.2 in debug mode. These issues
seem to be resolved in Eigen's 'master' branch at time of writing this,
but are present in multiple recent releases that are supported by isce3.

The patch implements matrix multiplication as a sequence of dot
products. This implementation probably isn't optimal in general (and may
inhibit expression template chaining) but, in the common case of small,
fixed-size matrices/vectors, this implementation generates the same
assembly as before on nvcc 11.2 and works around the bug with nvcc 11.3.