Fix debug builds (#981)
Fix an issue that caused debug builds to fail due to an invalid
operation (<bool> && <std::string>) inside assert statements. The code
was modified to first stream debug info to a pyre::journal::firewall_t
channel prior to the assertion. These assertions are in vendored code
from ortools, anyway, so we don't expect to trip them in isce3.