Bump minimum pyre version to 1.11.2 (#1046)
* Bump minimum pyre version to 1.11.2

* Update OL8 environment

Bumping up our minimum supported pyre version to v1.11.2 indirectly
introduced a few incompatibilities in our conda-forge-based dev
environments. conda-forge only generated pyre binaries for the updated
recipe with their most recent build toolchain (gcc==12.1.0), which was
more recent than the maximum version of gcc that nvcc officially
supports as a host compiler.

This commit includes updates to the Oracle Linux 8 image configuration
to support building with the new pyre version.

* Update the CUDA version from 11.0 -> 11.7

Originally, the plan was to upgrade the CI system's CUDA version in
order to raise the ceiling on the maximum supported host compiler. It
turns out that this is insufficient -- the newest CUDA version (11.7)
still doesn't have official support for gcc>=12 -- so we had to
introduce some additional workarounds.

We could probably revert this change safely if we want to, but I've
decided to keep it for now as part of a general upgrade to the CI
environment.

* Bump maximum version of g++
* Remove version caps on libgcc and libstdc++

Previously, we had capped the versions of these compiler & runtime
packages to be <10 in order to satisfy maximum host compiler version
requirements of our CUDA distribution.

Now that the CUDA version has been upgraded to 11.7, g++ versions up to
v12 are officially supported as a host compiler.

Since libgcc and libstdc++ are backwards compatible, there was actually
no need to restrict the maximum versions of these libraries. These
version restrictions have now been removed

* Add a runtime dependency on sysroot_linux-64>=2.17

This was necessary to resolve link-time errors due to unresolved symbols
in glibc.

* Loosen `NVIDIA_REQUIRE_CUDA` version restrictions

This was causing docker builds to fail when the CUDA version inside the
image was greater than the host version. It seems like this check was
overly strict -- due to minor version compatibility changes introduced
in CUDA 11, any host-side driver version that supports CUDA 11 should
work with any container-side CUDA version within that major version
family.

* Update Miniconda version from 4.9.2 -> 4.12.0

Like the upgraded CUDA version, this ended up not being strictly
necessary, but I didn't see any reason to revert the change. I
originally bumped the version to try and fix a conda bug that I ran into
(https://github.com/conda/conda/issues/9672).