Fix InSAR workflow test runconfigs (#1056)
The 3 InSAR workflow test cases have become out-of-sync with the
current runconfig schema. A while back, the 'blocksize' group was
removed from the schema and replaced with a 'lines_per_block' parameter
in the 'geo2rdr' and 'rdr2geo' groups. This update fixes the workflow
test runconfig yaml files to resolve schema validation errors.