Downgrade CUDA in OL8 images to 11.4 (#1104)
This change is to address some issues that the PGE and PCM teams ran
into when testing the R3.1 release image. Both teams encountered the
following error when trying to run a GPU-enabled SAS workflow on a
p3.2xlarge instance with CUDA<11.7 installed on the host:

```
std::bad_alloc: cudaErrorUnsupportedPtxVersion: the provided PTX was
compiled with an unsupported toolchain
```

CUDA applications, unlike most other applications, are not completely
isolated from the host environment when run in docker containers. The
host must have a version of the Nvidia display driver installed that is
compatible with the version of CUDA that the software was compiled
against.

For CUDA>=11 applications, drivers from older CUDA versions within the
same major version family are nominally supported by default
(https://docs.nvidia.com/deploy/cuda-compatibility/#default-to-minor-version).
This means that there is some degree of backwards compatibility with
older drivers that was not possible in previous CUDA versions. In
addition, applications compiled with older CUDA versions have always
been compatible with newer drivers due to binary backward compatibility
(even before CUDA 11).

It turns out that there are limitations to the same-major-version
backwards compatibility with older drivers, however. For one thing,
applications that are compiled to PTX, but not all the way to
architecture-specific cubin, are not compatible with older driver
versions
(https://docs.nvidia.com/deploy/cuda-compatibility/#application-considerations).

This limitation appears to have been the cause of the issues experienced
by the PGE & PCM teams. In our last few releases, we bumped the CUDA
version within the docker image to 11.7. Our docker scripts generate PTX
and compile to cubin for any native GPU architecture(s) found on the
build host. (This is the default behavior of isce3's CMake scripts and
our docker scripts don't appear to be overriding it.)

Therefore, our GPU-enabled software was benefitting from backwards
compatibility with older drivers *only* when run on GPUs that had the
exact same compute capability as the GPU(s) found on the build host. In
all other cases, the host's display driver needed to be at least as
recent as CUDA 11.7.

So when the PGE/PCM team tried to take our applications that had been
built on a p2.xlarge with CUDA 11.7 and run them on a p3.2xlarge with
CUDA 11.4 installed on the host, the applications failed at runtime.

Susan Neely has reported that display drivers newer than CUDA 11.4 are
unsupported by the K80 GPUs in AWS p2.xlarge instances. Therefore our
release docker builds should continue to support at least CUDA 11.4 for
the foreseeable future. All of our development, test, and production
instances have this driver or a newer driver installed on them.
Therefore, to resolve this issue, it is sufficient to downgrade our
CUDA version from 11.7 -> 11.4 for future releases.