Fix CI conda issue
The CI build is failing due to this conda issue:
https://github.com/conda/conda/issues/9672

Basically, we're installing a newer version of Python inside the base
environment which seems to be incompatible with conda itself. For the
time being, I've resolved this by pinning Python 3.9 (the version that
the conda installer ships with) and re-generating the spec files.

If/when we want to upgrade Python versions within the CI environment,
we'll need to fetch a newer version of the conda installer from
https://repo.anaconda.com/miniconda.