Mount SAS input dir in container during QA (#1340)
Currently, when we run workflow tests via docker, we mount the input,
output, and scratch directories inside the container. However, when we
run QA on the SAS product via docker, we only mount the output (and QA
output) directory inside the container.

This worked fine in previous releases, but recently the RSLC QA pipeline
was updated such that it may depend on some additional ancillary files,
such as a corner reflector CSV file or the input DEM file. This is
currently causing some issues in our docker-based testing, since these
files are stored in the *input* directory, which is not mounted in the
container during QA.

This update modifies the steps in `imgset.py` that run QA so that the
necessary inputs are mounted inside the container during QA processing.
The implementation is very similar to the workflow test steps that
already mount these input directories.