Reduce blocksize of GSLC UAVSAR test cases (#1345)
When running GSLC on UAVSAR data, a saw-tooth artifact is showing up
when running the workflow with large block sizes (e.g., 1000x1000) . The
problem has been only observed for UAVSAR data

The issue is related to the blocking function. Some geocoded blocks when
transformed to radar coordinates, should have data, but for some reason
they are deternined to not fall inside the data part. further analysis
needed to understand what exactly happens for those blocks.

Current workaround is to reduce the block size to 200x200.