Update RSLC test runconfigs (#1494)
* Disable EAP in RSLC test cases except for REE2

Disable EL antenna pattern correction in RSLC workflow test cases for
which we don't have an antenna_pattern file and/or
internal_calibration file. Both are required for this step to be run.

(The REE2 test case has both files so we keep EAP enabled for that test
case.)

* Change scale factor to 1.0

Change scale factor to 1.0 for REE point target simulations. The
previous scale factor was 1e-8, which caused RSLC QA to fail now that
the images are radiometrically calibrated since the image values were
all extremely small (close to machine epsilon).