Fix credentials in parameterized Jenkins jobs (#1547)
In Jenkins "Pipeline" jobs (like our "nisar-weekly-workflow-tests" and
"MERGE-workflow-mintests" jobs), credentials can be parameterized in the
job configuration directly via the Jenkins UI.

However, when these parameters are expanded during the pipeline
execution, they expand to the credential *ID* rather than the actual
secret value. This is causing the aforementioned CI jobs to fail.

We can resolve this by creating new environment variables to store the
credentials in the Jenkinsfiles, and expanding each parameter into the
`credentials()` operator in order to avoid exposing the credential IDs
in the codebase.