Add workaround for Docker permission errors due to Herbie (#1589)
Herbie (a transitive dependency via RAiDER) checks for a config file in
`~/.config/herbie/config.toml` when it's first imported and creates the
file if it did not exist. This tends to cause permission errors during
typical usage of NISAR ADT Docker images since generally users will not
have a username or home directory inside the container image (`$HOME`
expands to `/`).

We don't actually use Herbie in our workflows (but don't have any
control over whether or not it gets incidentally imported by RAiDER). As
a bandaid solution until the issue is addressed upstream, we will write
a default Herbie config file to the expected location during the `docker
build` step in order to prevent Herbie from trying to create it at
runtime.