Loosen dtype restrictions on SNAPHU inputs & outputs (#1785)
SNAPHU only allows single-precision floating-point datatypes, and only
supports 8-bit or 32-bit unsigned integer datatypes for the output
connected components labels.

However, this isn't really a limitation on the datatypes that the Python
wrapper can support in practice. We always copy the input data from
rasters to flat binary files, and likewise copy SNAPHU's outputs from
flat binary files to the output rasters. We can therefore convert the
dtype when copying, if necessary.

Previously, the Python wrapper for SNAPHU implemented overly-restrictive
checks that forced the input/output raster datatypes to exactly match
the datatypes required by SNAPHU. This caused issues for the InSAR
workflow now that we've changed the datatype of the connected components
layer to uint16.

These checks have now been loosened to remove restrictions on the
bit width of the input/output datatypes -- except for the mask layer,
which must still be an 8-bit integer type.