Revert changes from #1781 (#1794)
Reverts the changes from #1781.

That PR introduced a change to how the InSAR SAS determines the radar band ("L" or "S") of the input reference RSLC, which caused a breakage in backwards compatibility with (much) older products that don't have a `radarBand` field in their `identification` group (see #1791).

There is an ongoing effort by the UAVSAR team to update NISAR sample products to conform to the most recent spec. However, this is expected to take a while. Meanwhile, the outdated products are used in our own CI testing and widely disseminated in the community. Therefore, we decided to reintroduce a bandaid in the InSAR processor that allows it to process nonconforming products in the meantime.

The changes in this PR should exactly undo the changes from #1781. I created the PR using:

```
$ git diff 5d5caac..5d5caac^ --binary > patch
$ git apply patch
```

(where `5d5caac` is the SHA of the commit that merged #1781 into develop)