Disable SET corrections in GSLC UAVSAR SanAnd test case (#1880)
Disable solid earth tide (SET) corrections in one of the older GSLC test
cases. There's a known issue with this dataset -- it contains a raster
file with an invalid EPSG code that causes topo to fail. We plan to
replace this test dataset (and others) with a more up-to-date test case
in the future, but for now let's silence this error by disabling SET
corrections.