Bump minor version
This corresponds to a change in the way developmental versions of isce3
will be labeled. Previously, we would bump the minor version number in
the develop branch at the *end* of each release cycle, such that the
version number of "-dev" versions of isce3 lagged behind the version
number of the upcoming release.

This could cause confusion when comparing developmental versions to
release versions, especially following patch releases, because the patch
release number was ahead of the "-dev" version, even if the git history
of the patch was *behind* the develop branch.

Going forward, we will increment the minor version number in the develop
branch at the *beginning* of the release cycle, so that the version
number of developmental versions is always ahead of all prior
major/minor/patch releases.

This new versioning policy is also more consistent with PEP 440
conventions:

https://peps.python.org/pep-0440/#developmental-releases