Add radar observation time to PTA output (#1985)
* Add radar observation time to PTA output

Adds a new field "timestamp" to the JSON output of the RSLC & GSLC PTA
tools. This field stores the UTC datetime that the target was observed
by the radar in zero-Doppler geometry. This field was already present in
the AbsCal tool output but I omitted it from the PTA tool results for
some reason (even though we're already calculating it when we compute
the elevation angle of the corner reflector).

I see from
[MCR-88014](https://wiki.jpl.nasa.gov/pages/resumedraft.action?draftId=934290853&draftShareId=cf6d4b50-1131-4860-8eab-b2bf6ea8f768&#)
that the NISAR FOT expects the radar observation datetime of each corner
reflector to be present in the RSLC QA report. This PR will facilitate
adding that information to the QA output.

* Add type annotations in docstring for JSON fields

* Document datetime string precision